+++ src/main/java/com/yzhh/backstage/api/dao/mapper/BaseMapper.java
@@ -5,25 +5,25 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface BaseMapper <T, Texample> {
-	long countByExample(Texample example);
+	Long countByExample(Texample example);
 
-    int deleteByExample(Texample example);
+	Long deleteByExample(Texample example);
 
-    int deleteByPrimaryKey(Integer id);
+	Long deleteByPrimaryKey(Long id);
 
     List<T> selectByExample(Texample example);
 
-    int updateByExample(@Param("record") T record, @Param("example") Texample example);
+    Long updateByExample(@Param("record") T record, @Param("example") Texample example);
 
-    int updateByPrimaryKey(T record);
+    Long updateByPrimaryKey(T record);
     
-    int insert(T record);
+    Long insert(T record);
 
-    int insertSelective(T record);
+    Long insertSelective(T record);
 
-    T selectByPrimaryKey(Integer id);
+    T selectByPrimaryKey(Long id);
 
-    int updateByExampleSelective(@Param("record") T record, @Param("example") Texample example);
+    Long updateByExampleSelective(@Param("record") T record, @Param("example") Texample example);
 
-    int updateByPrimaryKeySelective(T record);
+    Long updateByPrimaryKeySelective(T record);
 }
